Output 2 from GPT: ##################################### Response from GPT:
Upon reviewing your error message, it seems a SyntaxError occurred. SyntaxError typically happens when Python parser fails to understand a line of code. The error usually contains information about where the error occurred at and what caused it. However, the error message that was provided is not clear as it doesn't provide any specific details about the problem in the code.

However, upon reviewing the provided code, it seems there is no obvious syntax error. The exception being the usage of DataLoader function on TRAIN_DATA_PATH and LABEL_FILE in line 61 without enclosing this with ICUData class. Please, see the corrected code below.