Output 2:
Error from previous code:
  File "<stdin>", line 19, in get_gpt_resp_code
  File "<string>", line 51, in <module>
  File "<string>", line 32, in __init__
<class 'TypeError'>
super(type, obj): obj must be an instance or subtype of type

Response from GPT:
Apologies for the mistake. The issue lies in the LSTM class definition. When calling `super()` in the initializer method of the LSTM, there is no object passed to `super().__init__()` when that class is subclassing a class which requires initialization. It should be just `super(LSTM, self).__init__()` instead of `super(self).__init__()`.

Below is the corrected code:
**********