Output 4:
Error from previous code:
  File "<stdin>", line 19, in get_gpt_resp_code
  File "<string>", line 51, in <module>
  File "<string>", line 32, in __init__
<class 'TypeError'>
super(type, obj): obj must be an instance or subtype of type

Response from GPT:
I apologize for the misunderstanding. The error message indicates that there exists a problem with the super() call in the LSTM class initiation. The correct way to call the super() function in Python3 is super().__init__(). Here is the complete, corrected code: