Output 6:
Error from previous code:
  File "<stdin>", line 19, in get_gpt_resp_code
  File "<string>", line 58, in <module>
  File "/data/sls/scratch/pschro/anaconda3/envs/env310/lib/python3.10/site-packages/torch/utils/data/dataloader.py", line 628, in __next__
    data = self._next_data()
  File "/data/sls/scratch/pschro/anaconda3/envs/env310/lib/python3.10/site-packages/torch/utils/data/dataloader.py", line 671, in _next_data
    data = self._dataset_fetcher.fetch(index)  # may raise StopIteration
  File "/data/sls/scratch/pschro/anaconda3/envs/env310/lib/python3.10/site-packages/torch/utils/data/_utils/fetch.py", line 58, in fetch
    data = [self.dataset[idx] for idx in possibly_batched_index]
  File "/data/sls/scratch/pschro/anaconda3/envs/env310/lib/python3.10/site-packages/torch/utils/data/_utils/fetch.py", line 58, in <listcomp>
    data = [self.dataset[idx] for idx in possibly_batched_index]
  File "<string>", line 25, in __getitem__
<class 'ValueError'>
could not convert string to float: 'Spontaneously'

Response from GPT:
I apologize for this issue. This error describes that data cannot be converted from string to float. One reason may be some non-numeric values like 'Spontaneously' are present in the numerical columns we are trying to convert.

The key here is to remove any non-numeric records or transform them into a numerical form that the model can understand. In this case, though, if 'Spontaneously' or other such non-numeric categories are not significant or cannot be transformed into a meaningful numerical representation, we should drop these kinds of records.

Here is the corrected code: