Output 7:
Error from previous code:
  File "<stdin>", line 19, in get_gpt_resp_code
  File "<string>", line 67, in <module>
  File "/data/sls/scratch/pschro/anaconda3/envs/env310/lib/python3.10/site-packages/torch/utils/data/dataloader.py", line 628, in __next__
    data = self._next_data()
  File "/data/sls/scratch/pschro/anaconda3/envs/env310/lib/python3.10/site-packages/torch/utils/data/dataloader.py", line 671, in _next_data
    data = self._dataset_fetcher.fetch(index)  # may raise StopIteration
  File "/data/sls/scratch/pschro/anaconda3/envs/env310/lib/python3.10/site-packages/torch/utils/data/_utils/fetch.py", line 58, in fetch
    data = [self.dataset[idx] for idx in possibly_batched_index]
  File "/data/sls/scratch/pschro/anaconda3/envs/env310/lib/python3.10/site-packages/torch/utils/data/_utils/fetch.py", line 58, in <listcomp>
    data = [self.dataset[idx] for idx in possibly_batched_index]
  File "<string>", line 32, in __getitem__
<class 'NameError'>
name 'is_float' is not defined

Response from GPT:
I apologize for the oversight. The error indicates that the name 'is_float' is not defined in the `__getitem__` function. I did not properly scope the helper function `is_float`. It was defined outside the dataset class, which is why it was not recognized within the class. I will now define it as a static method in the ICUData class. Here's the corrected code: