Output 4:
Error from previous code:
  File "<stdin>", line 22, in get_gpt_resp_code
  File "<string>", line 59, in <module>
  File "<string>", line 33, in __init__
<class 'NameError'>
name 'LSTMModel' is not defined

Response from GPT:
It seems like there is still an issue with the ordering of the code. The `device` object needs to be defined before the `LSTMModel` class because it's needed in the `forward` method of `LSTMModel`. Additionally, the lines of code that rely on the `LSTMModel` namely its instantiation as `model` and the following lines should come only after the `LSTMModel` class is defined. Here is the corrected code: