Output 7:
Error from previous code:
  File "<stdin>", line 22, in get_gpt_resp_code
  File "<string>", line 53, in <module>
  File "<string>", line 31, in __init__
<class 'NameError'>
name 'LSTMModel' is not defined

Response from GPT:
Apologies for the previous confusion. It seems like the initialization of the model is happening before the LSTMModel class is completely defined. Let's define the LSTM class & function first and then initialize the model. This should resolve the issue. Here is the revised version: