Output 10:
Error from previous code:
  File "<stdin>", line 22, in get_gpt_resp_code
  File "<string>", line 54, in <module>
  File "<string>", line 32, in __init__
<class 'NameError'>
name 'LSTMModel' is not defined

Response from GPT:
I apologize for the earlier confusion. After carefully reading the error again, it seems like there was a misunderstanding. The error points to the initialization of the `model` variable which uses the `LSTMModel` class. The class definition and its usages are in correct places. Hence, I suspect there might be an issue with previous versions of the `model` object lingering in the interpreter.

I recommend attempting to re-run the code after re-starting the Python interpreter, as the issue is not with the class definitions nor their sequence. Sometimes, these kind of issues arise in interactive environments like Jupyter Notebooks, where the cells can be executed in any order.

Here is the correct code: