Output 16:
Error from previous code:
  File "<stdin>", line 22, in get_gpt_resp_code
  File "<string>", line 69, in <module>
  File "/data/sls/scratch/pschro/anaconda3/envs/env310/lib/python3.10/site-packages/torch/utils/data/dataloader.py", line 628, in __next__
    data = self._next_data()
  File "/data/sls/scratch/pschro/anaconda3/envs/env310/lib/python3.10/site-packages/torch/utils/data/dataloader.py", line 671, in _next_data
    data = self._dataset_fetcher.fetch(index)  # may raise StopIteration
  File "/data/sls/scratch/pschro/anaconda3/envs/env310/lib/python3.10/site-packages/torch/utils/data/_utils/fetch.py", line 61, in fetch
    return self.collate_fn(data)
  File "<string>", line 33, in pad_collate
  File "<string>", line 33, in <listcomp>
  File "/data/sls/scratch/pschro/anaconda3/envs/env310/lib/python3.10/site-packages/torch/_tensor.py", line 894, in __len__
    raise TypeError("len() of a 0-d tensor")
<class 'TypeError'>
len() of a 0-d tensor

Response from GPT:
I apologize for the confusion from the previous responses. The error message arises because the function `len()` is called on a 0-d tensor. This suggests that we shouldn't try to compute the length of the labels tensor, as it's a scalar.

Since the data is already ordered and those with missing values are filled with zeros, a simple solution would be to pad all sequences to match the length of the sequence with the maximum length. Here is the the corrected code: