Output 2: Response to NameError
Error from previous code:
  File "<stdin>", line 23, in get_gpt_resp_code
  File "<string>", line 55, in <module>
  File "<string>", line 15, in __init__
<class 'NameError'>
name 'pd' is not defined

Response from GPT:
It seems the issue is related to the pandas library not being imported. But in the provided code, pandas is imported as pd at the beginning. It's possible that there may be other issues related to python environments, packages, or coding errors. Nonetheless, I'll provide a similar program with error corrections and better handling, please look through and let me know if you need any changes, or if you have further problems occurring.