Output 3: Response to TypeError
Error from previous code:
  File "<stdin>", line 2, in <module>
  File "<string>", line 90, in <module>
  File "<string>", line 55, in train
  File "/data/sls/scratch/pschro/anaconda3/envs/env310/lib/python3.10/site-packages/torch/utils/data/dataloader.py", line 628, in __next__
    data = self._next_data()
  File "/data/sls/scratch/pschro/anaconda3/envs/env310/lib/python3.10/site-packages/torch/utils/data/dataloader.py", line 671, in _next_data
    data = self._dataset_fetcher.fetch(index)  # may raise StopIteration
  File "/data/sls/scratch/pschro/anaconda3/envs/env310/lib/python3.10/site-packages/torch/utils/data/_utils/fetch.py", line 61, in fetch
    return self.collate_fn(data)
  File "<string>", line 34, in collate_fn
<class 'TypeError'>
expected Tensor as element 0 in argument 0, but got int

Response from GPT:
Based on the error message, the problem is associated with the data type of the output in the `collate_fn` function. This error is likely due to the incorrect usage of the `torch.stack` function with integer numbers. Note that `torch.stack` is expecting its inputs to be tensors.

The integer numbers (lengths of the patients' records) that you are trying to stack are not `torch.Tensor` types, and this is causing the TypeError. The solution is to convert the sequences' lengths to `torch.tensor` before stacking them in the `collate_fn` function.

Here's the corrected code: